requirements.txt
django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.0
psycopg2-binary==2.9.7
djangorestframework_simplejwt==5.3.1
celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.2
django-redis==5.4.0
